spanify: Handle `subspan()` args for containers
This CL is a basic extension of https://crrev.com/c/6482947 done for
container-ish types, e.g. `std::vector` or `std::array`.

Bug: 412395846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6487427
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1453731}

+++ tools/clang/spanify/tests/pointer-into-vector-expected.cc
@@ -4,10 +4,12 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // found in the LICENSE file.
 
 #include <array>
+#include <cstdint>
 #include <string>
 #include <vector>
 
 #include "base/containers/span.h"
+#include "base/numerics/safe_conversions.h"
 
 int UnsafeIndex();
 
@@ -33,17 +35,21 @@ int main() {
   TakePointerIntoContainer(vector);
 
   // Expected rewrite:
-  // TakePointerIntoContainer(base::span<int>(vector).subspan(2));
-  TakePointerIntoContainer(base::span<int>(vector).subspan(2));
+  // TakePointerIntoContainer(base::span<int>(vector).subspan(2u));
+  TakePointerIntoContainer(base::span<int>(vector).subspan(2u));
 
   int cached_index = UnsafeIndex();
   // Expected rewrite:
-  // TakePointerIntoContainer(base::span<int>(vector).subspan(cached_index));
-  TakePointerIntoContainer(base::span<int>(vector).subspan(cached_index));
+  // TakePointerIntoContainer(base::span<int>(vector).subspan(
+  //     base::checked_cast<size_t>(cached_index)));
+  TakePointerIntoContainer(base::span<int>(vector).subspan(
+      base::checked_cast<size_t>(cached_index)));
 
   // Expected rewrite:
-  // TakePointerIntoContainer(base::span<int>(vector).subspan(UnsafeIndex()));
-  TakePointerIntoContainer(base::span<int>(vector).subspan(UnsafeIndex()));
+  // TakePointerIntoContainer(base::span<int>(vector).subspan(
+  //     base::checked_cast<size_t>(UnsafeIndex())));
+  TakePointerIntoContainer(base::span<int>(vector).subspan(
+      base::checked_cast<size_t>(UnsafeIndex())));
 
   // Also test basic rewrite functionality on `std::array`.
   auto array = std::to_array<int>({13, 26, 39, 52});
@@ -53,8 +59,8 @@ int main() {
   TakePointerIntoContainer(array);
 
   // Expected rewrite:
-  // TakePointerIntoContainer(base::span<int>(array).subspan(2));
-  TakePointerIntoContainer(base::span<int>(array).subspan(2));
+  // TakePointerIntoContainer(base::span<int>(array).subspan(2u));
+  TakePointerIntoContainer(base::span<int>(array).subspan(2u));
 
   // TODO: no rewriting is done here. Investigate.
   // Test basic rewrite functionality on `std::string`.
